server: move metadata cleanup into server implementation
