+++ server/server.go
@@ -17,6 +17,7 @@ package server
 import (
 	"bytes"
 	"context"
+	"encoding/json"
 	"fmt"
 	"math"
 	"math/rand/v2"
@@ -2246,6 +2247,147 @@ func (s *Server) IsTTLConfigExist(key string) bool {
 		}
 	}
 	return false
+}
+
+var (
+	// ErrMicroserviceMetadataCleanupRejected indicates that the persisted
+	// keyspace-group state cannot be safely cleaned up.
+	ErrMicroserviceMetadataCleanupRejected = errors.New("microservice metadata cleanup rejected")
+	// ErrMicroserviceMetadataCleanupUnavailable indicates that the request could
+	// not be completed under the current normal-PD leadership term.
+	ErrMicroserviceMetadataCleanupUnavailable = errors.New("microservice metadata cleanup unavailable")
+)
+
+type microserviceMetadataCleanupTerm struct {
+	leaderKey   string
+	leaderValue string
+	leaseID     clientv3.LeaseID
+}
+
+func rejectMicroserviceMetadataCleanup(format string, args ...any) error {
+	return errors.Wrapf(ErrMicroserviceMetadataCleanupRejected, format, args...)
+}
+
+func unavailableMicroserviceMetadataCleanup(format string, args ...any) error {
+	return errors.Wrapf(ErrMicroserviceMetadataCleanupUnavailable, format, args...)
+}
+
+// CleanupMicroserviceMetadata clears the persisted Members field of the default
+// TSO keyspace group in normal PD mode and reports whether it changed. A nil
+// error is fenced by one exact normal-PD leadership term and represents a
+// linearizable check that the default group existed, had no transition in
+// progress, had no non-default sibling, and had no persisted member at that
+// point. The caller must ensure keyspace assignments have already been merged
+// into the default group and all microservice metadata writers are stopped.
+func (s *Server) CleanupMicroserviceMetadata(ctx context.Context) (bool, error) {
+	term, err := s.captureMicroserviceMetadataCleanupTerm()
+	if err != nil {
+		return false, err
+	}
+	if s.IsKeyspaceGroupEnabled() {
+		return false, rejectMicroserviceMetadataCleanup("pd is already running in microservice mode")
+	}
+
+	groupKey := keypath.KeyspaceGroupIDPath(constant.DefaultKeyspaceGroupID)
+	groupPrefixEnd := clientv3.GetPrefixRangeEnd(keypath.KeyspaceGroupIDPrefix())
+	operationCtx, cancel := context.WithTimeout(ctx, etcdutil.DefaultRequestTimeout)
+	defer cancel()
+	resp, err := s.client.Get(
+		operationCtx,
+		groupKey,
+		clientv3.WithRange(groupPrefixEnd),
+		clientv3.WithLimit(2),
+	)
+	if err != nil {
+		return false, unavailableMicroserviceMetadataCleanup(
+			"failed to read keyspace-group metadata: %v", err)
+	}
+
+	if len(resp.Kvs) == 0 {
+		return false, rejectMicroserviceMetadataCleanup("default TSO keyspace group does not exist")
+	}
+	if string(resp.Kvs[0].Key) != groupKey || len(resp.Kvs) > 1 {
+		return false, rejectMicroserviceMetadataCleanup("found a non-default TSO keyspace group")
+	}
+
+	group := &endpoint.KeyspaceGroup{}
+	if err := json.Unmarshal(resp.Kvs[0].Value, group); err != nil {
+		return false, errs.ErrJSONUnmarshal.Wrap(err).GenWithStackByCause()
+	}
+	if group.ID != constant.DefaultKeyspaceGroupID {
+		return false, rejectMicroserviceMetadataCleanup(
+			"found TSO keyspace group %d at the default group path", group.ID)
+	}
+	if group.IsSplitting() {
+		return false, rejectMicroserviceMetadataCleanup("default TSO keyspace group is splitting")
+	}
+	if group.IsMerging() {
+		return false, rejectMicroserviceMetadataCleanup("default TSO keyspace group is merging")
+	}
+	groupRevision := clientv3.Compare(
+		clientv3.ModRevision(groupKey),
+		"=",
+		resp.Kvs[0].ModRevision,
+	)
+	changed := len(group.Members) > 0
+
+	nonDefaultGroupStart := keypath.KeyspaceGroupIDPath(constant.DefaultKeyspaceGroupID + 1)
+	comparisons := []clientv3.Cmp{
+		clientv3.Compare(clientv3.Value(term.leaderKey), "=", term.leaderValue),
+		clientv3.Compare(clientv3.LeaseValue(term.leaderKey), "=", term.leaseID),
+		groupRevision,
+		clientv3.Compare(
+			clientv3.CreateRevision(nonDefaultGroupStart).WithRange(groupPrefixEnd),
+			"=",
+			0,
+		),
+	}
+	operation := clientv3.OpGet(groupKey)
+	if changed {
+		group.Members = nil
+		value, err := json.Marshal(group)
+		if err != nil {
+			return false, errs.ErrJSONMarshal.Wrap(err).GenWithStackByCause()
+		}
+		operation = clientv3.OpPut(groupKey, string(value))
+	}
+
+	failpoint.InjectCall("beforeCleanupMicroserviceMetadataCommit")
+	txnResp, err := kv.NewSlowLogTxnWithContext(operationCtx, s.client).
+		If(comparisons...).
+		Then(operation).
+		Commit()
+	if err != nil {
+		return false, unavailableMicroserviceMetadataCleanup(
+			"failed to commit microservice metadata cleanup: %v", err)
+	}
+	if txnResp.Succeeded {
+		return changed, nil
+	}
+	return false, unavailableMicroserviceMetadataCleanup(
+		"leadership or keyspace-group metadata changed during cleanup")
+}
+
+func (s *Server) captureMicroserviceMetadataCleanupTerm() (microserviceMetadataCleanupTerm, error) {
+	if s.client == nil || s.member == nil || !s.member.IsServing() {
+		return microserviceMetadataCleanupTerm{}, unavailableMicroserviceMetadataCleanup(
+			"normal PD leader is not serving")
+	}
+	leadership := s.member.GetLeadership()
+	if leadership == nil || leadership.GetLease() == nil {
+		return microserviceMetadataCleanupTerm{}, unavailableMicroserviceMetadataCleanup(
+			"normal PD leadership is not initialized")
+	}
+	term := microserviceMetadataCleanupTerm{
+		leaderKey:   leadership.GetLeaderKey(),
+		leaderValue: leadership.GetLeaderValue(),
+		leaseID:     leadership.GetLease().GetID(),
+	}
+	if term.leaderKey == "" || term.leaderValue == "" || term.leaseID == 0 {
+		return microserviceMetadataCleanupTerm{}, unavailableMicroserviceMetadataCleanup(
+			"normal PD leadership term is incomplete")
+	}
+	return term, nil
 }
 
 // MarkSnapshotRecovering mark pd that we're recovering
